# docs/SJewls_Admin_API_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SJewls Admin & Staff Portal API",
    "description": "API endpoints for the SJewls Admin Web Portal \u2014 Staff Authentication, Password Recovery, Staff User Management, Customer Management with Full Unmasked NIC, and Branch Statistics.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5230",
      "description": "Current Host"
    },
    {
      "url": "http://localhost:5230",
      "description": "Local Server"
    }
  ],
  "paths": {
...
    "/api/v1/admin/customers": {
      "get": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "List Customers with Pagination and Filters",
        "description": "Retrieves a paginated list of customers. Supports search by name, phone, email, or NIC, as well as filtering by branch and account status. Super Admins can access all branches; Branch staff are strictly restricted to their assigned branch.",
        "operationId": "AdminGetCustomers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerPagedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Create Customer by Staff",
        "description": "Allows authorized branch staff or Super Admin to create a customer. Normalizes inputs, prevents duplicate NICs and contacts, and keeps contacts unverified until customer verifies via OTP.",
        "operationId": "AdminCreateCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/admin/customers/statistics": {
      "get": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Customer Summary Card Statistics",
        "description": "Calculates customer counts (Total, Active, Inactive) across matching branch and search criteria prior to pagination. Status filter is not applied so active and inactive counts remain independent.",
        "operationId": "AdminGetCustomerStatistics",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerStatisticsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/admin/customers/{id}": {
      "get": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Get Detailed Customer Profile",
        "description": "Retrieves detailed profile information for an authorized customer, including date of birth, unmasked NIC for authorized staff, verification flags, and list of associated Chitu and Jewellery savings slots. Secrets, OTPs, and password hashes are never exposed.",
        "operationId": "AdminGetCustomerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/admin/customers/{id}/status": {
      "patch": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Activate or Deactivate Customer Account",
        "description": "Enables authorized staff to deactivate or reactivate a customer account within permitted branch scope. Deactivation requires a reason and immediately revokes all sessions/refresh tokens.",
        "operationId": "AdminUpdateCustomerStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminCustomerDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "nic": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "isPhoneVerified": {
            "type": "boolean"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "phoneVerifiedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "emailVerifiedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "primaryBranchId": {
            "type": "string",
            "format": "uuid"
          },
          "primaryBranchName": {
            "type": "string",
            "nullable": true
          },
          "primaryBranchCode": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "isProfileComplete": {
            "type": "boolean"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "deactivatedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivationReason": {
            "type": "string",
            "nullable": true
          },
          "closedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "closureReason": {
            "type": "string",
            "nullable": true
          },
          "chituSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCustomerSlotDto"
            },
            "nullable": true
          },
          "jewelleryEnrolments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCustomerEnrolmentDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AdminCustomerEnrolmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "planName": {
            "type": "string",
            "nullable": true
          },
          "planCode": {
            "type": "string",
            "nullable": true
          },
          "targetWeightGrams": {
            "type": "number",
            "format": "double"
          },
          "totalSavedGrams": {
            "type": "number",
            "format": "double"
          },
          "totalContributedCurrency": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "targetEndDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "AdminCustomerListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "nic": {
            "type": "string",
            "nullable": true
          },
          "isPhoneVerified": {
            "type": "boolean"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "primaryBranchId": {
            "type": "string",
            "format": "uuid"
          },
          "primaryBranchName": {
            "type": "string",
            "nullable": true
          },
          "primaryBranchCode": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "isProfileComplete": {
            "type": "boolean"
          },
          "activePlansCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalSlotsCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "deactivatedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivationReason": {
            "type": "string",
            "nullable": true
          },
          "closedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "closureReason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AdminCustomerPagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCustomerListItemDto"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "AdminCustomerSlotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slotNumber": {
            "type": "integer",
            "format": "int32"
          },
          "planName": {
            "type": "string",
            "nullable": true
          },
          "planCode": {
            "type": "string",
            "nullable": true
          },
          "monthlyInstalmentAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "activatedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateCustomerRequest": {
        "required": [
          "email",
          "fullName",
          "nic",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "maxLength": 150,
            "minLength": 2,
            "type": "string"
          },
          "phoneNumber": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "nic": {
            "minLength": 1,
            "type": "string"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomerStatisticsDto": {
        "type": "object",
        "properties": {
          "totalCustomers": {
            "type": "integer",
            "format": "int32"
          },
          "activeCustomers": {
            "type": "integer",
            "format": "int32"
          },
          "inactiveCustomers": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCustomerStatusRequest": {
        "required": [
          "isActive"
        ],
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT Bearer token (e.g., Bearer eyJhbGci...)",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}